Module orml_utilities::iterator

👎 Deprecated since 0.4.1:
iterator module’s functionality is now available in substrate’s frame-support

Structs

MapIteratorDeprecated

Iterate over a prefix and decode raw_key and raw_value into T. Forks from substrate, expose previous_key field

MapIteratorShimDeprecated

Shim for MapIterator, add more features

StorageMapIteratorDeprecated

Utility to iterate through items in a storage map. Forks from substrate, expose previous_key field

StorageMapIteratorShimDeprecated

Shim for StorageMapIterator, add more features

Traits

IterableStorageDoubleMapExtendedDeprecated

A strongly-typed map in storage whose keys and values can be iterated over.

IterableStorageMapExtendedDeprecated

A strongly-typed map in storage whose keys and values can be iterated over.